turn browse to stateless widget cause there's no reason for it to be stateful, add hint to the fontsize text field to show current set font size.

+++ lib/models/settings.dart
@@ -2,14 +2,14 @@ import 'package:flutter/cupertino.dart';
 import 'package:just_read/services/pdf_reader.dart';
 
 class Settings extends ChangeNotifier {
-  double _defaultFontSize = 20;
-  double _fontSize = 20;
+  int _defaultFontSize = 20;
+  int _fontSize = 20;
   PDFReader _pdfReader;
 
-  double get fontSize => _fontSize;
+  int get fontSize => _fontSize;
   PDFReader get pdf => _pdfReader;
 
-  void changeFont(double newFontSize) {
+  void changeFont(int newFontSize) {
     _fontSize = newFontSize > 10 ? newFontSize : _defaultFontSize;
     notifyListeners();
   }
